+++ ui/Models/OverlayTextAnchor.cs
@@ -1,0 +1,10 @@
+namespace Ca.Jwsm.Railroader.Api.Ui.Models
+{
+    public enum OverlayTextAnchor
+    {
+        TopLeft = 0,
+        TopRight = 1,
+        BottomLeft = 2,
+        BottomRight = 3,
+    }
+}